add rclone config

+++ install.sh
@@ -1,20 +1,26 @@
 #!/bin/bash
+set -e
 
-# 安装目录
+# installation script for dotfiles
 DOTFILES_DIR=$(pwd)
+# assert DOTFILES_DIR is dotfiles
+if [ ! -f "$DOTFILES_DIR/install.sh" ]; then
+  echo "Error: This script must be run from the dotfiles directory."
+  exit 1
+fi
 
-# 1. 配置 XDG 目录
+# 1. Configure XDG base directories
 XDG_CONFIG_HOME="${XDG_CONFIG_HOME:-$HOME/.config}"
 XDG_DATA_HOME="${XDG_DATA_HOME:-$HOME/.local/share}"
 
-# 2. 创建 XDG 目录（如果不存在）
+# 2. Create XDG directories (if they don't exist)
 mkdir -p "$XDG_CONFIG_HOME"
 mkdir -p "$XDG_DATA_HOME"
 
-# 3. 创建目标目录并创建软链接
-
+# 3. Create target directories and create symbolic links
 # shell
-export OSH="$XDG_CONFIG_HOME/oh-my-bash"; bash -c "$(curl -fsSL https://raw.githubusercontent.com/ohmybash/oh-my-bash/master/tools/install.sh)"
+# use omarchy's bash settings
+git clone --depth=1 https://github.com/basecamp/omarchy.git "$XDG_DATA_HOME/omarchy"
 ln -sf "$DOTFILES_DIR/bash_profile" "$HOME/.bash_profile"
 ln -sf "$DOTFILES_DIR/bashrc" "$HOME/.bashrc"
 ln -sf "$DOTFILES_DIR/profile" "$HOME/.profile"
@@ -31,10 +37,6 @@ ln -sf "$DOTFILES_DIR/gdb/gdbinit" "$XDG_CONFIG_HOME/gdb/gdbinit"
 mkdir -p "$XDG_CONFIG_HOME/maven"
 ln -sf "$DOTFILES_DIR/maven/settings.xml" "$XDG_CONFIG_HOME/maven/settings.xml"
 
-# fish
-mkdir -p "$XDG_CONFIG_HOME/fish"
-ln -sf "$DOTFILES_DIR/fish/config.fish" "$XDG_CONFIG_HOME/fish/config.fish"
-
 # fonts
 mkdir -p "$XDG_CONFIG_HOME/fonts"
 ln -sf "$DOTFILES_DIR/fonts/fonts.conf" "$XDG_CONFIG_HOME/fontconfig/fonts.conf"
@@ -46,13 +48,6 @@ ln -sf "$DOTFILES_DIR/environment.d/general-env.conf" "$XDG_CONFIG_HOME/environm
 # git
 mkdir -p "$XDG_CONFIG_HOME/git"
 ln -sf "$DOTFILES_DIR/git/config" "$XDG_CONFIG_HOME/git/config"
-
-# hypr
-mkdir -p "$XDG_CONFIG_HOME/hypr/scripts"
-ln -sf "$DOTFILES_DIR/hypr/hyprland.conf" "$XDG_CONFIG_HOME/hypr/hyprland.conf"
-ln -sf "$DOTFILES_DIR/hypr/hyprpaper.conf" "$XDG_CONFIG_HOME/hypr/hyprpaper.conf"
-ln -sf "$DOTFILES_DIR/hypr/scripts/backlight" "$XDG_CONFIG_HOME/hypr/scripts/backlight"
-ln -sf "$DOTFILES_DIR/hypr/scripts/volume" "$XDG_CONFIG_HOME/hypr/scripts/volume"
 
 # kitty
 mkdir -p "$XDG_CONFIG_HOME/kitty"
@@ -87,14 +82,6 @@ popd
 pushd $VIM_PLUGIN_OPT_HOME
 git clone --depth=1 https://github.com/neoclide/coc.nvim.git
 popd
-
-# nvim
-ln -sf "$DOTFILES_DIR/nvim" "$XDG_CONFIG_HOME/nvim"
-
-# waybar
-mkdir -p "$XDG_CONFIG_HOME/waybar"
-ln -sf "$DOTFILES_DIR/waybar/config" "$XDG_CONFIG_HOME/waybar/config"
-ln -sf "$DOTFILES_DIR/waybar/style.css" "$XDG_CONFIG_HOME/waybar/style.css"
 
 # vscode
 mkdir -p "$XDG_CONFIG_HOME/Code/User"
